% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickets_summarized.R
\docType{data}
\name{tickets_summarized}
\alias{tickets_summarized}
\title{Summarized ticket data from {fakir}}
\format{
A data frame with 200 rows and 7 variables
}
\source{
\url{https://github.com/ThinkR-open/fakir}
}
\usage{
tickets_summarized
}
\description{
This is randomly generated data using the `{fakir}` package. This data is summarized from the `tickets` data to reduce computing the summarized data for the datatable output.
}
\keyword{datasets}
